Fixed minor format string warning in i_prologix.c
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1387 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/i_prologix.c
@@ -472,7 +472,7 @@ mxi_prologix_write( MX_GPIB *gpib,
 			"The supplied buffer is too long to fit into the "
 			"%lu character write buffer for the Prologix GPIB "
 			"interface '%s'.",
-				prologix->write_buffer_length,
+				(unsigned long) prologix->write_buffer_length,
 				gpib->record->name );
 		}
 
@@ -496,7 +496,7 @@ mxi_prologix_write( MX_GPIB *gpib,
 				"The supplied buffer is too long to fit into "
 				"the %lu character write buffer for the "
 				"Prologix GPIB interface '%s'.",
-					prologix->write_buffer_length,
+				(unsigned long) prologix->write_buffer_length,
 					gpib->record->name );
 			}
